==> scRNA-Seq-invivo/analysis_pipeline/notebooks/9_paper_figure.Rmd <==
---
title: Ageing & Wnt in an organoid model of the mouse small instenial stem cell niche - analysis
author: "Jan Gleixner, Maja Funk"
date: "`r Sys.Date()`"
params:
  aggregated_data_dir: "results_ci_human/analysis_in_vivo/3_aggregation"
  processed_data_dir: "results_ci_human/analysis_in_vivo/2_cell_type_assignment"
  differential_expression_dir: "results_ci_human/analysis_in_vivo/4_differential_expression"
  results_dir: "results"
---

```{r do setup, code=readLines(fs::path(rprojroot::find_rstudio_root_file(), "notebooks", "setup_chunk.R")), echo=FALSE, eval=TRUE}
source(file = fs::path(rprojroot::find_rstudio_root_file(), "notebooks", "setup_chunk.R")) # for RStudio
```


``` {r, warn=FALSE, message=FALSE, echo=FALSE}
library(sir)
library(schelpr)

library(data.table)
library(tidyverse)
library(patchwork)
library(colorspace)
```

```{r}
dt <- fread(fs::path(processed_data_dir, "meta_and_reductions.tsv"), sep= "\t")
```

```{r tSNE_by_sample_id}
dt %>% ggplot(aes(x=tSNE_1, y=tSNE_2, color=sample_id)) +
  geom_cell() +
  # scale_color_sample_id() +
  NULL
```

```{r tSNE_by_cell_type}
dt %>% ggplot(aes(x=tSNE_1, y=tSNE_2, color=cell_type)) +
  geom_cell() +
  scale_color_cell_type() +
  NULL
```

```{r tSNE_by_age}
dt %>% ggplot(aes(x=tSNE_1, y=tSNE_2, color=condition)) +
  geom_cell() +
  scale_color_age() +
  NULL
```


```{r inflammaging score}
dt %>% ggplot(aes(x=tSNE_1, y=tSNE_2, color=nCount_RNA)) + 
  geom_cell() + 
  scale_color_continuous(trans=schelpr::make_quantile_trans(dt[, nCount_RNA]))+
  NULL

dt %>% ggplot(aes(x=tSNE_1, y=tSNE_2, color=inflammaging_score)) + 
  geom_cell() + 
  scale_color_continuous(trans=schelpr::make_quantile_trans(dt[, inflammaging_score]))+
  NULL

dt %>% ggplot(aes(x=tSNE_1, y=tSNE_2, color=inflammaging_score_scaled)) + 
  geom_cell() + 
  scale_color_continuous(trans=schelpr::make_quantile_trans(dt[,inflammaging_score_scaled]))+
  NULL

dts <- dt[, .(inflammaging_score = mean(inflammaging_score), inflammaging_score_scaled = mean(inflammaging_score_scaled)), by = .(cell_type, condition, Tx_run_number, sample_id)]
dts %>% ggplot(aes(x= inflammaging_score, y=  cell_type, color = condition)) + stat_summary() +scale_color_age() + scale_y_cell_type()
dts %>% ggplot(aes(x= inflammaging_score_scaled, y=  cell_type, color = condition)) + stat_summary() +scale_color_age() + scale_y_cell_type()
```


```{r inflammaging score sina , fig.width=12}
dt %>% ggplot(aes(y= inflammaging_score_scaled, x=  cell_type, color = condition, group = interaction(cell_type,sample_id))) + ggforce::geom_sina() +scale_color_age() + scale_x_cell_type() +
  theme(aspect.ratio = 1/5)
dt %>% ggplot(aes(y= inflammaging_score_scaled, x=  cell_type, color = condition, split=as.integer(as.factor(condition)), group = interaction(cell_type,sample_id))) +
  geom_split_sina( position=position_dodge(width=0.8), scale = "width", maxwidth=0.8) +
  scale_color_age() +
  scale_x_cell_type() +
  theme(aspect.ratio = 1/5) +
  NULL
```


```{r inflammaging score Enterocytes }
dt[cell_type %in% "Enterocyte"] %>% ggplot(aes(x= inflammaging_score_scaled, color = condition, linetype = as.factor(Tx_run_number))) +
  geom_density() +
  scale_color_age()

dt[cell_type %in% "Enterocyte"][, mean_inflammaging_score := mean(inflammaging_score_scaled), by= sample_id][, inflammaging_score_batch_adj:= inflammaging_score_scaled-mean(unique(mean_inflammaging_score)), by = Tx_run_number] %>% ggplot(aes(x= inflammaging_score_batch_adj, color = condition, linetype = as.factor(Tx_run_number))) +
  geom_density() +
  scale_color_age()

```



```{r}
n_cells_dt <- dt[, .(n_cells = .N), by = .(cell_type, sample_id, treatment, condition)][, fraction_of_cells := n_cells/sum(n_cells), by = sample_id]

n_cells_dt[, mean_fraction_of_cells := mean(fraction_of_cells), by= cell_type][, rare := mean_fraction_of_cells < max(mean_fraction_of_cells)/4]

dodge_width <- 0.9
point_jitter_width <- 0.0
this_dodge <- function(width = dodge_width, reverse = TRUE, padding = 0.2) position_dodge2(width = width, reverse = reverse, padding = padding)

p <- n_cells_dt %>% ggplot(aes(x = cell_type, y = fraction_of_cells, color = condition)) +
  stat_summary(aes(fill = condition), fun.data = mean_se, geom = "col", position = this_dodge(), width = dodge_width) +
  stat_summary(aes(color = stage(condition, after_scale = colorspace::darken(color, 0.3))), fun.data = mean_se, geom = "errorbar", position = this_dodge(padding=1/3), width = dodge_width) +
  stat_summary(aes(group = interaction(cell_type, condition), color = stage(condition, after_scale = colorspace::darken(color, 0.3))), fun = mean, geom = "crossbar", position = this_dodge(), width = dodge_width) +
  geom_point(position = position_jitterdodge(point_jitter_width, dodge.width = dodge_width), aes(group = interaction(cell_type, forcats::fct_rev(condition)), fill=condition), color = "black") +
  labs(y = "fraction of cells per sample") +
  guides(color = "none") +
  scale_y_continuous(labels = scales::percent_format(accuracy = 1)) +
  scale_color_age() +
  scale_fill_age() +
  scale_x_cell_type() +
  theme(axis.title.x = element_blank()) +
  NULL
p

p + ggforce::facet_zoom(ylim = c(0, 0.1))
p + ggforce::facet_zoom(ylim = c(0, 0.1), zoom.size = 1)



q <- p + facet_wrap("rare", scales = "free") + theme(aspect.ratio = NULL, strip.background = element_blank(), strip.text = element_blank())
q
```

```{r coposition-split}
library(grid)
gt = ggplotGrob(q)

N <- n_cells_dt[, .(N = length(unique(cell_type))), by = rare][, N]

# Get the column index in the gt layout corresponding to the panels.
panelI <- gt$layout$l[grepl("panel", gt$layout$name)]

# Replace the default panel widths with relative heights.
gt$widths[panelI] <- unit(N, "null")

# Add extra width between panels (assuming two panels)
gt$widths[panelI[1] + 1] = unit(1, "cm")

## Draw gt
grid.newpage()
grid.draw(gt)
```


```{r, error = TRUE}
n_cells_dt[, broom::tidy(t.test(log(fraction_of_cells)~condition, data = .SD)), by = cell_type][, .(relative_change_fraction_ci = pmax(1-exp(conf.low), exp(conf.high)-1)), by = cell_type][]
n_cells_dt[, broom::tidy(t.test(fraction_of_cells~condition, data = .SD)), by = cell_type
           ][, .(mean_fraction_young = estimate2, absolute_change_fraction_ci = pmax(abs(conf.low), abs(conf.high)), conf.low, conf.high, pval = p.value), by = cell_type
             ][, relative_change_fraction_ci := scales::percent_format(1)(absolute_change_fraction_ci/mean_fraction_young)
               ][, adj_pval := p.adjust(pval, method = "BH")][]
```


# DEGs

```{r}
de_single_cell_dt <- fread(fs::path(differential_expression_dir, "DE_samples_as_replicates.tsv"), sep="\t")
de_single_cell_dt <- de_single_cell_dt[contrast == "condition_young_vs_aged"]
de_single_cell_dt[, gene_name := gene]
setnames(de_single_cell_dt, 
         old =  c("estimate", "stderror", "statistic", "p.value", "p.adjusted"),
         new = c("log2FoldChange", "lfcSE", "stat", "pvalue", "padj")
)

cell_types <- de_single_cell_dt[, sort(unique(cell_type))]
cell_types <- c(cell_types[cell_types %in% "pseudo bulk"], cell_types[!(cell_types %in% "pseudo bulk")])

de_single_cell_dt[is.na(padj), padj := 1]
```


```{r visualize number of degs}
de_single_cell_dt %>% ggplot(aes(x = cell_type, y = -log2FoldChange)) +
  ggrastr::rasterize(geom_jitter(color = "gray"), dpi = 600) +
  #geom_jitter(aes(color = cell_type), data = de_single_cell_dt[padj<0.1]) +
  ggforce::geom_sina(aes(color = cell_type), data = de_single_cell_dt[padj<0.1], scale="width") +
  #ggforce::geom_sina()
  labs(y = "log2fc aged/young") +
  guides(color = "none") +
  scale_x_cell_type() +
  scale_color_cell_type() +
  theme(axis.title.x = element_blank()) +
  NULL
```


```{r volcano, fig.height=3.5}
n_highlight <- 20
volcano_plot(de_single_cell_dt[cell_type=="pseudo bulk",],
  highlight = rank(- (abs(log2FoldChange)-(lfcSE*qnorm(1-0.1*n_highlight/length(lfcSE)))))<=n_highlight, 
  label_size = 2.5, 
  force_pull = 5, 
  point.padding = 0.2,
  force = 70,
  #nudge_y = 10,
  color = "black",     # text color
  bg.color = "white", # shadow color
  bg.r = 0.2          # shadow radius
) +
  facet_wrap("cell_type") +
  NULL
```
```{r volcano with core genes, fig.height=3.5}
number_of_top_cell_types <- 5
most_abundant_cell_types <- n_cells_dt[, .(n_cells =sum(n_cells)), by=.(cell_type)][order(-n_cells)]
most_abundant_cell_types
core_cell_types <- most_abundant_cell_types[seq_len(number_of_top_cell_types), cell_type]
core_cell_types
core_cell_types <- setdiff(core_cell_types, "Goblet")
core_cell_types
core_genes <- de_single_cell_dt[padj <=0.1][cell_type %in% core_cell_types][, .(n_cell_types = .N), by = .(gene)][n_cell_types == length(core_cell_types), gene]
core_genes

volcano_plot(de_single_cell_dt[cell_type=="pseudo bulk",],
  highlight = gene %in% core_genes, 
  label_size = 2.5, 
  force_pull = 5, 
  point.padding = 0.2,
  force = 70,
  #nudge_y = 10,
  color = "black",     # text color
  bg.color = "white", # shadow color
  bg.r = 0.2          # shadow radius
) +
  facet_wrap("cell_type") +
  NULL
```

```{r mhcII pval}
mhcII_features_dt <- fread("SI_feature_gene_lists.tsv")[feature %in% "MHC class II"]
mhcII_features_dt <- de_single_cell_dt[mhcII_features_dt, on = .(gene = marker_gene_symbol)]
effect_range <- mhcII_features_dt[, max(abs(log2FoldChange))] * c(-1, 1)
symmetric_limits <- function(lim) max(abs(lim)) * c(-1, 1)
```

```{r, fcr, fig.height=3.5}
alpha <- 0.1
mhcII_features_dt[, log2fc_lower := log2FoldChange + lfcSE * qnorm(alpha/2*pvalue/padj)]
mhcII_features_dt[, log2fc_upper := log2FoldChange + lfcSE * qnorm(alpha/2*pvalue/padj, lower.tail = FALSE)]

mhcII_features_dt %>%
  ggplot(aes(x = gene, y = -log2FoldChange, color = cell_type)) +
  geom_point(position = position_dodge2(width = 0.8)) +
  geom_errorbar(aes(ymin = -log2fc_lower, ymax = -log2fc_upper), position = position_dodge2(width = 0.8, padding = 0.2), width=0.8) +
  scale_y_continuous("log2fc aged/young") +
  scale_color_cell_type() +
  guides(x = guide_axis(angle=45)) +
  theme(aspect.ratio = 1/2) +
  NULL
```


```{r, nominal-ci, fig.height=3.5}
alpha <- 0.1
mhcII_features_dt[, log2fc_lower := log2FoldChange + lfcSE * qnorm(alpha/2)]
mhcII_features_dt[, log2fc_upper := log2FoldChange + lfcSE * qnorm(alpha/2, lower.tail = FALSE)]

p <- mhcII_features_dt %>%
  ggplot(aes(x = gene, y = -log2FoldChange, color = cell_type, group = factor(cell_type, levels = sir:::levels_of_cell_type, ordered = TRUE))) +
  geom_point(position = position_dodge2(width = 0.8)) +
  geom_errorbar(aes(ymin = -log2fc_lower, ymax = -log2fc_upper), position = position_dodge2(width = 0.8, padding = 0.2), width=0.8) +
  scale_y_continuous("log2fc aged/young") +
  scale_color_cell_type() +
  guides(x = guide_axis(angle=45)) +
  theme(aspect.ratio = 1/2) +
  NULL

p
p + coord_cartesian(ylim = mhcII_features_dt[, range(-log2FoldChange)] * (1.1^c(-1, 1)))

```


```{r mhcII pval 1, fig.height=3.5, fig.width=7}
r <- as(hex2RGB(sir:::maja_red),"polarLUV")@coords
b <- as(hex2RGB(sir:::maja_blue),"polarLUV")@coords
scale_color_effect_size <- function(name = "log2fc aged\nvs. young", c1 = 200, cmax = 200, l1 = 50, l2=10, h1 = b[,3], h2 = r[,3], n_interp = 51, limits = symmetric_limits, ...) 
                                    colorspace::scale_color_continuous_diverging(name = name, c1 = c1, cmax = cmax, l1 = l1, l2 = l2, h1 = h1, h2 = h2, n_interp = n_interp, limits = limits, ...)

mhcII_features_dt %>%
  ggplot(aes(x = gene, y = cell_type, color = -log2FoldChange)) +
  geom_point(aes(size = -log10(padj))) +
  #theme(panel.background = element_rect(fill = 'gray')) +
  scale_size_continuous("negative log10\nadjusted p-value")+
  scale_y_cell_type() +
  #theme(axis.line = element_blank()) +
  scale_color_effect_size() +
  guides(fill = guide_colorbar(), x=guide_axis(angle=90)) +
  theme(legend.direction = "vertical", legend.box = "horizontal") +
  NULL
```


```{r, fig.height=3.5}
mhcII_features_dt %>%
  ggplot(aes(x = gene, y = cell_type, color = -log2FoldChange)) +
  geom_point(aes(size = lfcSE)) +
  scale_size_continuous(trans='reverse')+
  scale_y_cell_type() +
  scale_color_effect_size() +
  guides(fill = guide_colorbar(), x=guide_axis(angle=90)) +
  theme(legend.direction = "vertical", legend.box = "horizontal") +
  NULL
```



```{r, supplementary figure qc filtering}
dt %>% ggplot(aes(x=nCount_RNA, y=fraction_mito, color=sample_id)) +
  geom_cell() +
  #gghighlight::gghighlight(label_key = cell_type) +
  facet_wrap("cell_type", nrow=2) +
  labs(x = "counts", y = "fraction of counts from mitochondrial genes") +
  theme(
    panel.grid.major.x = element_line(color="gray95", size = 0.3),
    panel.grid.minor.x = element_line(color="gray95", size = 0.1),
    panel.grid.major.y = element_line(color="gray95", size = 0.3),
    panel.grid.minor.y = element_line(color="gray95", size = 0.1)
  ) +
  #theme_gray() + # to show minor ticks
  #annotation_logticks(outside=TRUE, sides = "b", short=unit(0.1, "char"), mid = unit(0.2, "char"), long= unit(0.3, "char")) +
  #theme(
  #  axis.ticks.length.x = unit(c(0.3,rep(0.1, 8)), "char"),
  #  axis.text.x = element_text(colour = c("black", rep(NA, 8)))
  #) +
  #coord_cartesian(clip = "off") +
  #schelpr::scale_x_log(breaks = schelpr::log_breaks(maj = FALSE, 10)) +
  schelpr::scale_x_log()+
  guides(color = guide_legend(override.aes = list(size = 1))) +
  NULL
```
